fix: image preview after select a result
